summary-service/requirements.txt
Django==5.0
djangorestframework==3.14.0
transformers==4.36.2
torch==2.5.1
python-dotenv==1.0.0
gunicorn==21.2.0
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
